+++ app/scripts/components/search-examples.ts
@@ -5,6 +5,7 @@ import statemachine from "@/statemachine"
 import { html } from "@/util"
 import settings from "@/settings"
 import { SearchExample } from "@/settings/app-settings.types"
+import { RootScope } from "@/root-scope.types"
 
 export default angular.module("korpApp").component("searchExamples", {
     template: html`
@@ -25,7 +26,7 @@ export default angular.module("korpApp").component("searchExamples", {
         "$rootScope",
         "$scope",
         "$location",
-        function ($rootScope, $scope: SearchExamplesScope, $location) {
+        function ($rootScope: RootScope, $scope: SearchExamplesScope, $location) {
             const $ctrl = this
 
             $scope.examples = undefined